feat: reorganize component structure; move Modal and Loading components to ui directory; update imports accordingly
