+++ src/components/index.ts
@@ -2,8 +2,8 @@ export * from './Navbar'
 export * from './FormComponents'
 export * from './Cards'
 export * from './SideMenu'
-export * from './Loading'
-export * from './Modal'
+export * from './ui/Loading'
+export * from './ui/Modal'
 export * from './ModalComponents'
 export * from './PatientComponents'
-export * from './ui/Fab'
+export * from './ui'
